# specs/003-frontend-backend-integration/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Humanoid Robotics RAG Chatbot API
  description: |
    REST API for the RAG-powered chatbot that serves the Physical AI & Humanoid Robotics textbook.
    Supports multi-turn conversations, selected-text queries, and session persistence.
  version: 1.0.0
  contact:
    name: API Support
    email: support@humanoid-robotics-book.com

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://your-backend.hf.space
    description: Production server (Hugging Face Spaces)

tags:
  - name: chat
    description: Chat endpoints for querying the textbook
  - name: sessions
    description: Session management for multi-turn conversations
  - name: health
    description: Health check endpoints

paths:
  /api/chat:
    post:
      tags:
        - chat
      summary: Send a chat query to the RAG agent
      description: |
        Submit a question about the textbook and receive a response from the RAG agent.
        Supports full-document retrieval or selected-text mode for contextual queries.
      operationId: chatQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              fullRetrieval:
                summary: Full retrieval query
                value:
                  query: "What is Gazebo simulation?"
                  session_id: null
                  selected_text: null
                  stream: false
              selectedText:
                summary: Selected-text query
                value:
                  query: "Explain this concept in more detail"
                  session_id: "550e8400-e29b-41d4-a716-446655440000"
                  selected_text: "Gazebo is an open-source 3D robotics simulator that allows developers to test robot algorithms in realistic environments."
                  stream: false
      responses:
        '200':
          description: Successful response from the agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                success:
                  summary: Successful query
                  value:
                    response: "Gazebo is an open-source 3D robotics simulator that provides realistic physics simulation, high-quality graphics, and sensor simulation for testing robot algorithms."
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    message_id: "660e8400-e29b-41d4-a716-446655440001"
                    retrieved_chunks: 5
                    response_time_ms: 2341.5
                    error: null
        '400':
          description: Invalid request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emptyQuery:
                  summary: Empty query
                  value:
                    detail: "Query cannot be empty or whitespace-only"
                    error_code: "INVALID_REQUEST"
        '422':
          description: Unprocessable entity (Pydantic validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  summary: Database connection error
                  value:
                    detail: "Database connection failed"
                    error_code: "INTERNAL_ERROR"
        '503':
          description: Service unavailable (LLM API or retrieval service down)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serviceUnavailable:
                  summary: LLM API unavailable
                  value:
                    detail: "LLM service temporarily unavailable"
                    error_code: "SERVICE_UNAVAILABLE"

  /api/chat/stream:
    post:
      tags:
        - chat
      summary: Send a chat query with streaming response
      description: |
        Submit a question and receive a streaming response for real-time token generation.
        Uses Server-Sent Events (SSE) protocol.
      operationId: chatQueryStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Streaming response (SSE)
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream with response tokens
              examples:
                streamingResponse:
                  summary: SSE stream example
                  value: |
                    data: {"token": "Gazebo", "done": false}

                    data: {"token": " is", "done": false}

                    data: {"token": " an", "done": false}

                    data: {"session_id": "550e8400-e29b-41d4-a716-446655440000", "message_id": "660e8400-e29b-41d4-a716-446655440001", "done": true}
        '400':
          $ref: '#/paths/~1api~1chat/post/responses/400'
        '422':
          $ref: '#/paths/~1api~1chat/post/responses/422'
        '500':
          $ref: '#/paths/~1api~1chat/post/responses/500'
        '503':
          $ref: '#/paths/~1api~1chat/post/responses/503'

  /api/sessions:
    post:
      tags:
        - sessions
      summary: Create a new chat session
      description: |
        Explicitly create a new session. Optional - sessions are auto-created if not provided in /api/chat.
      operationId: createSession
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  pattern: '^[a-zA-Z0-9\-]+$'
                  maxLength: 255
                  description: Optional user identifier (for future authentication)
                  example: "user-12345"
                metadata:
                  type: object
                  additionalProperties: true
                  description: Optional metadata (e.g., user agent, IP, feature flags)
                  example:
                    user_agent: "Mozilla/5.0"
                    source: "docusaurus"
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
              examples:
                newSession:
                  summary: New session created
                  value:
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    created_at: "2025-12-29T10:30:00Z"
                    last_accessed: "2025-12-29T10:30:00Z"
        '400':
          $ref: '#/paths/~1api~1chat/post/responses/400'
        '500':
          $ref: '#/paths/~1api~1chat/post/responses/500'

  /api/sessions/{session_id}:
    get:
      tags:
        - sessions
      summary: Retrieve session details and message history
      description: |
        Get all messages in a session for displaying conversation history.
      operationId: getSession
      parameters:
        - name: session_id
          in: path
          required: true
          description: UUID of the session
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        - name: limit
          in: query
          required: false
          description: Maximum number of messages to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          example: 20
        - name: offset
          in: query
          required: false
          description: Offset for pagination (skip N messages)
          schema:
            type: integer
            minimum: 0
            default: 0
          example: 0
      responses:
        '200':
          description: Session found with message history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetailResponse'
              examples:
                sessionWithMessages:
                  summary: Session with 2 messages
                  value:
                    session_id: "550e8400-e29b-41d4-a716-446655440000"
                    created_at: "2025-12-29T10:30:00Z"
                    last_accessed: "2025-12-29T10:35:00Z"
                    message_count: 2
                    messages:
                      - message_id: "660e8400-e29b-41d4-a716-446655440001"
                        role: "user"
                        content: "What is Gazebo simulation?"
                        selected_text: null
                        timestamp: "2025-12-29T10:30:15Z"
                        response_time_ms: null
                        error_message: null
                      - message_id: "660e8400-e29b-41d4-a716-446655440002"
                        role: "assistant"
                        content: "Gazebo is an open-source 3D robotics simulator..."
                        selected_text: null
                        timestamp: "2025-12-29T10:30:17Z"
                        response_time_ms: 2341.5
                        error_message: null
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: Session not found
                  value:
                    detail: "Session not found"
                    error_code: "NOT_FOUND"
        '500':
          $ref: '#/paths/~1api~1chat/post/responses/500'

    delete:
      tags:
        - sessions
      summary: Delete a session (soft delete)
      description: |
        Mark a session as archived. Messages are retained for analytics but not returned in queries.
      operationId: deleteSession
      parameters:
        - name: session_id
          in: path
          required: true
          description: UUID of the session
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: Session deleted successfully (no content)
        '404':
          $ref: '#/paths/~1api~1sessions~1{session_id}/get/responses/404'
        '500':
          $ref: '#/paths/~1api~1chat/post/responses/500'

  /health:
    get:
      tags:
        - health
      summary: Basic health check
      description: Returns 200 OK if the service is running
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2025-12-29T10:30:00Z"

  /api/health:
    get:
      tags:
        - health
      summary: Detailed health check
      description: |
        Returns health status of all dependencies (database, Qdrant, LLM API).
      operationId: apiHealthCheck
      responses:
        '200':
          description: All services healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
              examples:
                allHealthy:
                  summary: All services operational
                  value:
                    status: "healthy"
                    timestamp: "2025-12-29T10:30:00Z"
                    services:
                      database:
                        status: "healthy"
                        response_time_ms: 12.3
                      qdrant:
                        status: "healthy"
                        response_time_ms: 45.7
                      llm_api:
                        status: "healthy"
                        response_time_ms: 234.1
        '503':
          description: One or more services unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
              examples:
                databaseDown:
                  summary: Database connection failed
                  value:
                    status: "unhealthy"
                    timestamp: "2025-12-29T10:30:00Z"
                    services:
                      database:
                        status: "unhealthy"
                        error: "Connection timeout"
                      qdrant:
                        status: "healthy"
                        response_time_ms: 45.7
                      llm_api:
                        status: "healthy"
                        response_time_ms: 234.1

components:
  schemas:
    ChatRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
          description: User's question about the textbook
          example: "What is Gazebo simulation?"
        session_id:
          type: string
          format: uuid
          nullable: true
          description: Session ID for multi-turn conversations (omit for new session)
          example: "550e8400-e29b-41d4-a716-446655440000"
        selected_text:
          type: string
          maxLength: 5000
          nullable: true
          description: Highlighted text from textbook for contextual queries
          example: "Gazebo is an open-source 3D robotics simulator..."
        stream:
          type: boolean
          default: false
          description: Enable streaming response (use /api/chat/stream endpoint)

    ChatResponse:
      type: object
      required:
        - response
        - session_id
        - message_id
        - retrieved_chunks
        - response_time_ms
      properties:
        response:
          type: string
          description: Agent's response to the query
          example: "Gazebo is an open-source 3D robotics simulator that provides realistic physics simulation..."
        session_id:
          type: string
          format: uuid
          description: Session ID (newly created or provided)
          example: "550e8400-e29b-41d4-a716-446655440000"
        message_id:
          type: string
          format: uuid
          description: Unique message identifier
          example: "660e8400-e29b-41d4-a716-446655440001"
        retrieved_chunks:
          type: integer
          minimum: 0
          description: Number of textbook chunks retrieved (0 for selected-text mode)
          example: 5
        response_time_ms:
          type: number
          format: float
          minimum: 0
          description: Response time in milliseconds
          example: 2341.5
        error:
          type: string
          nullable: true
          description: Error message if generation failed (response may be partial)
          example: null

    SessionResponse:
      type: object
      required:
        - session_id
        - created_at
        - last_accessed
      properties:
        session_id:
          type: string
          format: uuid
          description: Unique session identifier
          example: "550e8400-e29b-41d4-a716-446655440000"
        created_at:
          type: string
          format: date-time
          description: Session creation timestamp
          example: "2025-12-29T10:30:00Z"
        last_accessed:
          type: string
          format: date-time
          description: Last activity timestamp
          example: "2025-12-29T10:30:00Z"

    SessionDetailResponse:
      allOf:
        - $ref: '#/components/schemas/SessionResponse'
        - type: object
          required:
            - message_count
            - messages
          properties:
            message_count:
              type: integer
              minimum: 0
              description: Total number of messages in session
              example: 2
            messages:
              type: array
              description: List of messages in chronological order
              items:
                $ref: '#/components/schemas/MessageDetail'

    MessageDetail:
      type: object
      required:
        - message_id
        - role
        - content
        - timestamp
      properties:
        message_id:
          type: string
          format: uuid
          description: Unique message identifier
          example: "660e8400-e29b-41d4-a716-446655440001"
        role:
          type: string
          enum: [user, assistant]
          description: Message sender role
          example: "user"
        content:
          type: string
          description: Message content
          example: "What is Gazebo simulation?"
        selected_text:
          type: string
          nullable: true
          description: Selected text for contextual queries (user messages only)
          example: null
        timestamp:
          type: string
          format: date-time
          description: Message timestamp
          example: "2025-12-29T10:30:15Z"
        response_time_ms:
          type: number
          format: float
          nullable: true
          description: Response time for assistant messages
          example: 2341.5
        error_message:
          type: string
          nullable: true
          description: Error message if generation failed
          example: null

    HealthCheckResponse:
      type: object
      required:
        - status
        - timestamp
        - services
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Overall health status
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: Health check timestamp
          example: "2025-12-29T10:30:00Z"
        services:
          type: object
          description: Health status of individual services
          properties:
            database:
              $ref: '#/components/schemas/ServiceHealth'
            qdrant:
              $ref: '#/components/schemas/ServiceHealth'
            llm_api:
              $ref: '#/components/schemas/ServiceHealth'

    ServiceHealth:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: Service health status
          example: "healthy"
        response_time_ms:
          type: number
          format: float
          nullable: true
          description: Response time in milliseconds (if healthy)
          example: 12.3
        error:
          type: string
          nullable: true
          description: Error message (if unhealthy)
          example: null

    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Query cannot be empty or whitespace-only"
        error_code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_REQUEST
            - NOT_FOUND
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - DATABASE_ERROR
            - LLM_ERROR
          example: "INVALID_REQUEST"

    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          description: List of validation errors from Pydantic
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Field location (e.g., ["body", "query"])
                example: ["body", "query"]
              msg:
                type: string
                description: Error message
                example: "field required"
              type:
                type: string
                description: Error type
                example: "value_error.missing"